# docs/openapi.yaml
openapi: 3.1.0
info:
  title: GELB CRM API
  version: 0.1.0
  description: API для MVP CRM языковой онлайн-школы
servers:
  - url: /api/v1
security:
  - cookieAuth: []
...
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Вход по логину и паролю
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [login, password]
              properties:
                login: { type: string }
                password: { type: string, format: password }
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUser'
        '401': { description: Неверные учетные данные }
...
  /teachers/{id}/access:
    get:
      tags: [Teachers]
      summary: Статус доступа преподавателя
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TeacherAccessStatus' }
        '404': { description: Teacher not found }
    post:
      tags: [Teachers]
      summary: Создать доступ преподавателя с временным паролем
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TeacherAccessCreated' }
        '409': { description: Access already exists or duplicate login }
        '422': { description: Teacher email required }
  /teachers/{id}/access/login:
    patch:
      tags: [Teachers]
      summary: Изменить логин доступа преподавателя
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [login]
              properties:
                login: { type: string, format: email }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  teacherId: { type: string, format: uuid }
                  login: { type: string, format: email }
        '404': { description: Teacher or access not found }
        '409': { description: Duplicate login }
  /teachers/{id}/access/password:
    patch:
      tags: [Teachers]
      summary: Изменить пароль доступа преподавателя и завершить активные сессии
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string, format: password }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  teacherId: { type: string, format: uuid }
                  forceLogoutAllSessions: { type: boolean }
        '404': { description: Teacher or access not found }
...
  /journal/weekly-template:
    get:
      tags: [Journal]
      summary: Недельный шаблон преподавателя
      parameters:
        - in: query
          name: teacherId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/WeeklySlot' }
    put:
      tags: [Journal]
      summary: Полная замена недельного шаблона
      parameters:
        - in: query
          name: teacherId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slots]
              properties:
                slots:
                  type: array
                  items: { $ref: '#/components/schemas/WeeklySlotUpsert' }
      responses:
        '204': { description: Updated }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: gelb_session
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
  schemas:
...
    AuthUser:
      type: object
      properties:
        id: { type: string, format: uuid }
        role: { type: string, enum: [admin, teacher] }
        login: { type: string }
...
    TeacherAccessStatus:
      type: object
      properties:
        teacherId: { type: string, format: uuid }
        hasAccess: { type: boolean }
        login: { type: string, format: email, nullable: true }
        lastLoginAt: { type: string, format: date-time, nullable: true }
    TeacherAccessCreated:
      type: object
      properties:
        teacherId: { type: string, format: uuid }
        login: { type: string, format: email }
        temporaryPassword: { type: string }
...
    WeeklySlot:
      type: object
      properties:
        id: { type: string, format: uuid }
        weekday: { type: integer, minimum: 1, maximum: 7 }
        startTime: { type: string, pattern: '^\\d{2}:\\d{2}$' }
        isActive: { type: boolean }
    WeeklySlotUpsert:
      type: object
      required: [weekday, startTime]
      properties:
        weekday: { type: integer, minimum: 1, maximum: 7 }
        startTime: { type: string, pattern: '^\\d{2}:\\d{2}$' }
        isActive: { type: boolean, default: true }
...